requirements.txt
flask==2.3.2
pandas==2.0.1
requests==2.31.0
snowflake-connector-python==3.0.3
python-dotenv==1.0.0